Equipment Database: Improved the Copy Item button
Now automatically saves the new copy when the button is pressed, and it appears on the master list to the left.
Necessitating this was additional code to handle what happens if the copy button is pressed and the auto generated new name is the same as an item that already exists.
Rather than creating a dialog box asking if you want to overwrite the existing item, it simply keeps generating new names until it finds one that isn't taken, and uses that as the name of the copy.
Improved the copy mechanism to properly fail the isDirty() test, which ensures you aren't bothered with "Discard Changes" dialog boxes when switching away from the edit view.